+++ CHANGELOG.rst
@@ -5,11 +5,13 @@ v3.2.1
 ------
 
 *Release date: In development*
+
 - Allow negative indexes for list elements in context searches. Example: [CONTEXT:element.-1]
 - Add support for JSON strings to the `DICT` and `LIST`` replacement. Example: [DICT:{"key": true}], [LIST:[null]]
 - Add `REPLACE` replacement, to replace a substring with another. Example: [REPLACE:[CONTEXT:some_url]::https::http]
 - Add `TITLE` replacement, to apply Python's title() function. Example: [TITLE:the title]
 - Add `ROUND` replacement, float number to a string with the indicated number of decimals. Example: [ROUND:3.3333::2]
+- Remove accents from generated file names to avoid errors in some filesystems
 
 v3.2.0
 ------